Add autodetection for Collabora and ONLYOFFICE in join script
I've been working on app appliances for Collabora and ONLYOFFICE the
last two weeks to finish them. Both apps have added a detection for
ownCloud and download or activate their respective apps in ownCloud,
when ownCloud is already installed.

During appliance setup, Collabora or ONLYOFFICE are installed first,
unfortunately. The test in the respective app's join script is false and
calls for ownCloud to active the apps are not executed. A forced
re-execution of the join script helps. But this is not the solution for
an appliance setup, where the user expects no more configuration work.

This detection is executed once upon installation of ownCloud. When
Collabora or ONLYOFFICE are already installed on the same host, the
respective apps in ownCloud are activated. This adds more comfort to the
owncloud app and produces benefit for the user.