added update checker option
with the default set to false to prevent the update notification when a new server release is available.

Updates should only be done via the univnention center, which has it's own update notifications. When a new app version is available then this notification get's triggered and a proper update can be performed.